pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "crowetrade"
version = "0.1.0"
description = "Crowe Logic Parallel Financial Agent Ecosystem — core skeleton"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [ { name = "Crowe Logic" } ]
dependencies = [
  "numpy>=1.24",
  "jsonschema>=4.21.0",
  "PyYAML>=6.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-cov>=4.1.0",
  "pytest-xdist>=3.5.0",
  "pytest-timeout>=2.2.0",
  "pytest-benchmark>=4.0.0",
  "hypothesis>=6.100.0",
  "ruff>=0.6.0",
  "black>=24.3.0",
  "mypy>=1.8.0",
  "jsonschema>=4.21.0",
  "pydantic>=2.5.0",
  "faker>=22.0.0",
]

[tool.pytest.ini_options]
addopts = [
  "-q",
  "--strict-markers",
  "--tb=short",
  "--cov=src/crowetrade",
  "--cov-report=term-missing:skip-covered",
  "--cov-report=html:htmlcov",
  "--cov-report=xml",
  "--cov-config=.coveragerc",
  "--cov-fail-under=5",
]
testpaths = ["tests"]
markers = [
  "unit: Unit tests",
  "contract: Contract tests",
  "integration: Integration tests",
  "e2e: End-to-end tests",
  "performance: Performance tests",
  "risk: Risk compliance tests",
  "simulation: Simulation/backtest tests",
  "slow: Slow running tests",
]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
lint.select = [
  "E","F","I","UP","B","A","C4","PIE","T20"
]
lint.ignore = [
  "E501"  # handled by black line-length
]

[tool.mypy]
python_version = "3.11"
pretty = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
strict_optional = true
no_implicit_optional = true
